issue #99 rename packages, add css, minor front changes (health page)
