+++ src/main/java/io/github/jhipster/registry/service/dto/ZuulRouteDTO.java
@@ -1,4 +1,4 @@
-package io.github.jhipster.registry.service.routeWrapper;
+package io.github.jhipster.registry.service.dto;
 
 import org.springframework.cloud.netflix.zuul.filters.ZuulProperties;
 
@@ -7,21 +7,21 @@ import java.util.Set;
 /**
  * Extends a ZuulRoute to add the instance status ("UP", "DOWN", etc...) .
  */
-public class RegistryRoute extends ZuulProperties.ZuulRoute {
+public class ZuulRouteDTO extends ZuulProperties.ZuulRoute {
 
     private String status;
 
-    public RegistryRoute(String id, String path, String serviceId, String url, boolean stripPrefix, Boolean retryable, Set<String> sensitiveHeaders, String status) {
+    public ZuulRouteDTO(String id, String path, String serviceId, String url, boolean stripPrefix, Boolean retryable, Set<String> sensitiveHeaders, String status) {
         super(id, path, serviceId, url, stripPrefix, retryable, sensitiveHeaders);
         this.status = status;
     }
 
-    public RegistryRoute(String path, String location, String status) {
+    public ZuulRouteDTO(String path, String location, String status) {
         super(path, location);
         this.status = status;
     }
 
-    public RegistryRoute(String status) {
+    public ZuulRouteDTO(String status) {
         super();
         this.status = status;
     }
