+++ src/main/webapp/app/admin/health/health.css
@@ -1,0 +1,16 @@
+/* ==========================================================================
+Health page styles
+========================================================================== */
+
+#health {
+    display: flex;
+    flex-flow: column;
+}
+
+span.label{ /* UP, DOWN, ... */
+    float: right;
+}
+
+li.list-group-item { /* routes names */
+    word-break: break-all;
+}
